+++ src/genie/entropy/jbig/encoder.h
@@ -66,11 +66,11 @@ class JBIGEncoder {
         two_line_template = jbigParameters.two_line_template;
     }
 
-    void encode(std::stringstream& input, std::stringstream& output, uint64_t ncols,
-        uint64_t nrows);
-    void encode(std::vector<uint8_t>& input, std::vector<uint8_t>& output, uint64_t ncols, uint64_t nrows);
+    void encode(std::stringstream& input, std::stringstream& output, uint32_t ncols,
+        uint32_t nrows);
+    void encode(std::vector<uint8_t>& input, std::vector<uint8_t>& output, uint32_t ncols, uint32_t nrows);
 
-    void decode(std::stringstream& input, std::stringstream& output, uint64_t& nrows, uint64_t& ncols);
+    void decode(std::stringstream& input, std::stringstream& output, uint32_t& nrows, uint32_t& ncols);
 
 
  private:
